+++ CHANGES.rst
@@ -117,6 +117,8 @@ Cubeviz
 - Fixes Spectral Extraction's assumptions of one data per viewer, and flux data only in
   flux-viewer/uncertainty data only in uncert-viewer. [#2646]
 
+- Fixed a bug in cube model fitting where multiprocessing fails when NaN is encountered. [#2682]
+
 Imviz
 ^^^^^